房间列表和编辑

+++ app/Core/Repository/AreaRepository.php
@@ -34,6 +34,12 @@ class AreaRepository extends BaseRepository
         return $this->code(200, "楼栋列表", $res);
     }
 
+
+    public function listArea()
+    {
+        // $columns=["area_no","area_name","area_size",'area_type','rental_price','rental_unit','orientations','renovation_type','layout_type',''];
+    }
+
     /**
      * 创建房间
      * @param array $param
@@ -41,29 +47,17 @@ class AreaRepository extends BaseRepository
      */
     public function createArea(array $param)
     {
-        $floor_no = (int)substr(sprintf("%04d", $param['area_no']), 0, 2);
-        $floor_chk = [
-            'build_sn' => $param['build_sn'],
-            'floor_no' => $floor_no,
-        ];
-        //1:检查楼层是否存在，若不存在则创建
-        $chk_floor_res = $this->floorService->first($floor_chk);
-        if ($chk_floor_res) {
-            $floor_sn = $chk_floor_res['floor_sn'];
-        } else {
-            $add_floor_res = $this->floorService->createFloor($param['build_sn'], $floor_no);
-            if (!$add_floor_res) {
-                return $this->code(400, "楼层创建失败");
-            }
-            $floor_sn = $add_floor_res['floor_sn'];
-        }
-        //2:检查房屋是否存在
-        $chk_area_res = $this->areaService->first(['area_sn' => $param['area_no']], ['area_no']);
+
+        //1:检查房屋是否存在
+        $chk_area_res = $this->areaService->first(['area_sn' => $param['area_sn']], ['area_no']);
         if ($chk_area_res) {
             return $this->code(400, "房间已存在");
         }
 
-        //2:创建房间
+        //获取楼层sn
+        $floor_sn = $this->getFloorSn($param['area_no'], $param['build_sn']);
+
+        //3:创建房间
         $area_res = $this->areaService->createArea($param['build_sn'], $floor_sn, $param['area_no'], $param['area_name'], $param['area_size'], $param['area_type'],
             $param['is_investment'], $param['orientations'], $param['rental_price'], $param['rental_unit'], $param['renovation_type'], $param['layout_type']
             , $param['bedroom_num'], $param['wc_room_num'], $param['drawing_room_num'], $param['introduce_imgs'], $param['layout_img'], $param['introduce_video']
@@ -77,14 +71,22 @@ class AreaRepository extends BaseRepository
     public function editArea(array $param)
     {
         //1:检查房屋是否存在
-        $chk_area_res = $this->areaService->first(['area_sn' => $param['area_no']], ['area_no']);
+        $chk_area_res = $this->areaService->first(['area_sn' => $param['area_sn']], ['lease_sn']);
         if (!$chk_area_res) {
             return $this->code(400, "房间不存在");
         }
+        //房间是否在租赁中
+        $is_rental = empty($chk_area_res['lease_sn']) ? false : true;
 
-        //2:检查房屋是否在租赁中
+        //修改房间数据
+        $area_res = $this->areaService->editArea($is_rental, $param['area_sn'], $param['area_name'], $param['area_size'], $param['area_type'],
+            $param['is_investment'], $param['orientations'], $param['rental_price'], $param['rental_unit'], $param['renovation_type'], $param['layout_type'],
+            $param['bedroom_num'], $param['wc_room_num'], $param['drawing_room_num'], $param['introduce_imgs'], $param['layout_img'], $param['introduce_video']);
+        if ($area_res) {
+            return $this->code(200, "房间编辑成功", ['area_sn' => $param['area_sn']]);
+        }
+        return $this->code(400, "房间编辑失败");
 
-        //3:修改房屋
 
     }
 
@@ -95,12 +97,14 @@ class AreaRepository extends BaseRepository
     public function deleteArea(array $param)
     {
         $where = ["area_sn" => $param['area_sn']];
-        $chk_res = $this->areaService->first($where);
+        $chk_res = $this->areaService->first($where, ['lease_sn']);
         if (!$chk_res) {
             return $this->code(400, "房间不存在");
         }
         //2:查询房间否在租赁当中
-
+        if (!empty($chk_res['lease_sn'])) {
+            return $this->code(400, "租赁中不能被删除");
+        }
         //3:执行删除操作
         $del_res = $this->areaService->deleteArea($param['area_sn']);
         if ($del_res) {
@@ -108,5 +112,32 @@ class AreaRepository extends BaseRepository
         }
         return $this->code(400, "房间删除失败");
 
+    }
+
+    /**
+     * 获取指定楼栋，房号所在的楼层sn
+     * @param string $area_no 房号
+     * @param string $build_sn 楼栋sn
+     * @return array|mixed
+     */
+    private function getFloorSn(string $area_no, string $build_sn)
+    {
+        $floor_no = (int)substr(sprintf("%04d", $area_no), 0, 2);
+        $floor_chk = [
+            'build_sn' => $build_sn,
+            'floor_no' => $floor_no,
+        ];
+        //1:检查楼层是否存在，若不存在则创建
+        $chk_floor_res = $this->floorService->first($floor_chk);
+        if ($chk_floor_res) {
+            $floor_sn = $chk_floor_res['floor_sn'];
+        } else {
+            $add_floor_res = $this->floorService->createFloor($build_sn, $floor_no);
+            if (!$add_floor_res) {
+                return $this->code(400, "楼层创建失败");
+            }
+            $floor_sn = $add_floor_res['floor_sn'];
+        }
+        return $floor_sn;
     }
 }
